% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateReplicates.R
\name{aggregateReplicates}
\alias{aggregateReplicates}
\title{Calculate point estimates for weights and bootstrap weights}
\usage{
aggregateReplicates(dat, FUN, groupVars = NULL, weights = attr(dat,
  "weights"), b.weights = attr(dat, "b.rep"))
}
\arguments{
\item{dat}{A \link{data.table} containing bootstrap weights. This will usually be the output of
\link{recalib} or \link{draw.bootstrap}.}

\item{FUN}{An aggregation function as returned by \link{WeightedMean}, \link{WeightedMedian}, \link{WeightedSum},
\link{WeightedRatio} and \link{WeightedSize}.}

\item{groupVars}{Column names in \code{dat} refering to grouping variables. Typically, those
columns are \code{factor}s or \code{integer}s}

\item{weights}{Column names for the survey weights}

\item{b.weights}{Column names for the boostrap weights}
}
\description{
Calculate an estimate for each group and each bootstrap replicate of survey weights. The
replicate weihgts are assumed to be in wide format as returned by \link{recalib} and
\link{draw.bootstrap}
}
\examples{
eusilc <- demo.eusilc(prettyNames = TRUE)
dat_boot <- draw.bootstrap(eusilc, REP = 10, hid = "hid", weights = "pWeight",
                           strata = "region", period = "year")
dat_boot_calib <- recalib(dat_boot, conP.var = c("gender", "age"), conH.var = c("region", "hsize"))

## poverty rate for each year
aggregateReplicates(dat_boot_calib, WeightedRatio("povertyRisk == 1"), c("year"))

## mean income for each year and gender
aggregateReplicates(dat_boot_calib, WeightedMean("eqIncome", na.rm = TRUE), c("year", "gender"))
}
