% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap.rep}
\alias{bootstrap.rep}
\title{bootstrap.rep}
\usage{
bootstrap.rep(dat, REP = 1000, hid = "hid", weights = "hgew",
  strata = "bundesld", year = "jahr", totals = NULL, boot.names = NULL)
}
\arguments{
\item{dat}{either data.frame or data.table containing the sample survey for various years.}

\item{REP}{integer indicating the number of bootstrap replicates.}

\item{hid}{string specifying the name of the column in \code{dat} containing the household ID.}

\item{weights}{string specifying the name of the column in \code{dat} containing the sample weights.}

\item{strata}{string vector specifying the name of the column in \code{dat} by which the population was stratified.}

\item{year}{string specifying the name of the column in \code{dat} containing the sample years.}

\item{totals}{string specifying the name of the column in \code{dat} containing the the totals per strata. If totals is NULL, the sum of weights per strata will be calcualted and named 'fpc'.}

\item{boot.names}{character indicating the leading string of the column names for each bootstrap replica. If NULL defaults to "w".}
}
\value{
the survey data with the number of REP bootstrap replicates added as columns.
}
\description{
Draw bootstrap replicates from survey data taken in various years.
Survey information, like ID, sample weights, strata and population totals per strata, should be specified to ensure meaningfull survey bootstraping.
}
\examples{

}
\author{
Johannes Gussenbauer, Statistik Austria
}
