% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.surveysd}
\alias{plot.surveysd}
\title{Title}
\usage{
\method{plot}{surveysd}(dat, variable = dat$param$var[1],
  type = c("summary", "grouping"), groups = NULL, sd.type = c("dot",
  "ribbon"))
}
\arguments{
\item{dat}{object of class 'surveysd' output of function \code{\link{calc.stError}}}

\item{variable}{Name of the variable for which standard errors have been calcualated in \code{dat}}

\item{type}{can bei either 'summary' or 'grouping', default value is 'summary'.
For 'summary' a barplot is created giving an overview of the number of estimates having the flag \code{smallGroup}, \code{cvHigh}, both or none of them.
For 'grouping' results for point estimate and standard error are plotted for pre defined groups.}

\item{groups}{If \code{type='grouping'} variables must be defined by which the data is grouped. Only 2 levels are supported as of right now.
If only one group is defined the higher group will be the estimate over the whole year.
Results are plotted for the first argument in \code{groups} as well as for the combination of \code{groups[1]} and \code{groups[2]}.}

\item{sd.type}{can bei either \code{'ribbon'} or \code{'dot'} and is only used if \code{type='grouping'}. Default is \code{"dot"}
For \code{sd.type='dot'} point estimates are plotted and flagged if the corresponding standard error and/or the standard error using the mean over k-years exceeded the value \code{cv.limit} (see \code{\link{calc.stError}}).
For \code{sd.type='ribbon'} the point estimates including ribbons, defined by point estimate +- estimated standard error are plotted.
The calculated standard errors using the mean over k years are plotted using less transparency. Results for the higher level (~\code{groups[1]}) are coloured grey.}
}
\description{
Title
}
