% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PointEstimates}
\alias{PointEstimates}
\alias{weightedRatio}
\alias{weightedSum}
\title{Weighted Point Estimates}
\usage{
weightedRatio(x, w)

weightedSum(x, w)
}
\arguments{
\item{x}{numeric vector}

\item{w}{weight vector}
}
\value{
Each of the functions return a single numeric value
}
\description{
Predefined functions for weighted point estimates in package \code{surveysd}.
}
\details{
Predefined functions are weighted ratio and weighted sum.
}
\examples{
x <- 1:10
w <- 10:1
weightedRatio(x,w)
x <- 1:10
w <- 10:1
weightedSum(x,w)
}
