% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stError.R
\name{print.surveysd}
\alias{print.surveysd}
\title{Print function for surveysd objects}
\usage{
\method{print}{surveysd}(x, ...)
}
\arguments{
\item{x}{an object of class \code{'surveysd'}}

\item{...}{additonal parameters}
}
\description{
Prints the results of a call to \code{\link[surveysd]{calc.stError}}. Shows used variables and function, number of point estiamtes
as well as properties of the results.
}
