% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ipf.R
\name{summary.ipf}
\alias{summary.ipf}
\title{Generate summary output for a ipf calibration}
\usage{
\method{summary}{ipf}(x, extraExport = NULL)
}
\arguments{
\item{x}{object of class ipf}

\item{extraExport}{additional exports to compute}
}
\value{
a list of the following outputs
}
\description{
Generate summary output for a ipf calibration
}
\examples{

\dontrun{
# load data
eusilc <- demo.eusilc(n = 1, prettyNames = TRUE)

# personal constraints
conP1 <- xtabs(pWeight ~ age, data = eusilc)
conP2 <- xtabs(pWeight ~ gender + region, data = eusilc)
conP3 <- xtabs(pWeight*eqIncome ~ gender, data = eusilc)

# household constraints
conH1 <- xtabs(pWeight ~ hsize + region, data = eusilc)

# simple usage ------------------------------------------

calibweights1 <- ipf(
  eusilc,
  conP = list(conP1, conP2, eqIncome = conP3),
  bound = NULL,
  verbose = TRUE
)
summary.ipf(calibweights1)
}
}
\author{
Laura Gruber
}
