% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateReplicates.R
\name{WeightedMean}
\alias{WeightedMean}
\alias{WeightedMedian}
\alias{WeightedSum}
\alias{WeightedRatio}
\alias{WeightedSize}
\title{Point estimate functions}
\usage{
WeightedMean(var, na.rm = FALSE)

WeightedMedian(var, na.rm = FALSE)

WeightedSum(var, na.rm = FALSE)

WeightedRatio(condition, na.rm = FALSE)

WeightedSize(condition, na.rm = FALSE)
}
\arguments{
\item{var}{A variable name in the dataset to be used for estimation}

\item{na.rm}{logical. Should missing values (including \code{NaN}) be removed?}

\item{condition}{A string specifying a condition that can be evaluated for each observation.
This condition may contain column names in the dataset.}
}
\value{
A function taking a subset of the data and a weight vector as input and returning
a single numeric value. For example \code{function(subset, w){ sum(w) }}
}
\description{
Helper functions to define point estimates for \link{aggregateReplicates}.
}
\details{
\code{WeightedRatio} returns the relative number of observations in each group
satisfying \code{condition}. For example, a poverty-risk rate for each state.
}
\examples{
df <- data.table::data.table(age = rnorm(100, mean = 30, sd = 10), w = 1)

estimator <- WeightedMean("age")
estimator(df, df$w)

estimator <- WeightedRatio("age>30")
estimator(df, df$w)


estimator <- WeightedSize("age>30")
estimator(df, df$w)
}
