% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stError.R
\name{calc.stError}
\alias{calc.stError}
\title{calc.stError}
\usage{
calc.stError(dat, weights = "hgew", b.weights = paste0("w", 1:1000),
  year = "jahr", var = "povmd60", fun = "weightedRatio",
  cross_var = list(NULL), year.diff = NULL, add.arg = vector("list",
  length(var)))
}
\arguments{
\item{dat}{either data.frame or data.table containing the sample survey for various years.}

\item{weights}{character specifying the name of the column in \code{dat} containing the sample weights. Used to calculate yearly and 3-year estimates (defined by \code{var} and \code{fun}).}

\item{b.weights}{character specifying the names of the columns in \code{dat} containing bootstrap weights. Used to calculate standard error of yearly and 3-year estimates (defined by \code{var} and \code{fun}).}

\item{year}{character specifying the name of the column in \code{dat} containing the sample years.}

\item{var}{character vector containing variable names in \code{dat} on which \code{fun} shall be applied for each \code{year}. Must have the same length as \code{fun}.}

\item{fun}{character vector containig function-names which are to be applied on \code{var} for each \code{year}. Must have the same length as \code{var}.
Possible arguments are \code{weightedRatio,weightedSum,sampSize,popSize} as well as any other function which returns a double or integer and has \code{weight} as its second argument.}

\item{cross_var}{list of character vectors containig variables in \code{dat}. For each list entry \code{dat} will be split in subgroups according to the containing variables as well as \code{year}
and \code{var} and \code{fun} are calculated for each of those groups.}

\item{year.diff}{character vectors, defining the years for which the differenzes as well as the standard deviation of the differenzes should be computed. Each entry must have the form of \code{"year1 - year2"}. Can be NULL}

\item{add.arg}{list containing strings for additional function arguments. Must be the same length as \code{var} oder \code{fun}. Can be a list of \code{NULL}s.}
}
\value{
data.table containing yearly and 3-year estimates of \code{fun} applied to \code{var} as well as the corresponding standard errors, which were calculated using the bootstrap weights.
}
\description{
Calcualte estimates of variables per year as well as their 3 year rolling mean.
In addition the standard deviation per year and for each three consecutive years are calculated using the bootstrap weights.
}
\author{
Johannes Gussenbauer, Statistik Austria
}
