% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{draw.bootstrap}
\alias{draw.bootstrap}
\title{Draw bootstrap replicates}
\usage{
draw.bootstrap(dat,REP=1000,hid="hid",weights="hgew",strata="bundesld",
                     year="jahr",totals=NULL,boot.names=NULL)
}
\arguments{
\item{dat}{either data.frame or data.table containing the survey data with rotating panel design.}

\item{REP}{integer indicating the number of bootstrap replicates.}

\item{hid}{character specifying the name of the column in \code{dat} containing the household ID.}

\item{weights}{character specifying the name of the column in \code{dat} containing the sample weights.}

\item{strata}{character vector specifying the name of the column in \code{dat} by which the population was stratified.}

\item{year}{character specifying the name of the column in \code{dat} containing the sample years.}

\item{country}{character specifying the name of the column in \code{dat} containing the country name. Is only used if \code{dat} contains data from multiple countries.
In this case the bootstep procedure will be applied on each country seperately. If \code{country=NULL} the household identifier must be unique for each household.}

\item{cluster}{character vector specifying cluster in the data. If \code{NULL} household ID is taken es the lowest level cluster.}

\item{totals}{(optional) character specifying the name of the column in \code{dat} containing the the totals per strata and/or cluster. If totals and cluster is \code{NULL}, the households per strata will be calcualted using the \code{weights} argument and named 'fpc'.
If clusters are specified then totals need to be supplied by the user, otherwise they will be set to \code{NULL}. When multiple cluster and or strata are specified totals needs to contain multiple argument each corresponding to a column name in \code{dat}.
Each column needs to contains the total number of units in the population regarding the subsequent level. The vector is interpreted from left to right meaning that the most left value of \code{totals}
specifies the column names with the number of units in the population at the highest level and the most right value specifies the column names with the number of units in the population at the lowest level.
This argument will be passed onto the function \code{svydesign()} from package \code{survey} through the argument \code{fpc}.}

\item{boot.names}{character indicating the leading string of the column names for each bootstrap replica. If NULL defaults to "w".}
}
\value{
the survey data with the number of REP bootstrap replicates added as columns.

Returns a data.table containing the original data as well as the number of \code{REP} columns containing the bootstrap replicates for each repetition.\cr
The columns of the bootstrap replicates are by default labeled "w\emph{Number}" where \emph{Number} goes from 1 to \code{REP}.
If the column names of the bootstrap replicates should start with a different character or string the parameter \code{boot.names} can be used.
}
\description{
Draw bootstrap replicates from survey data with rotating panel design.
Survey information, like ID, sample weights, strata and population totals per strata, should be specified to ensure meaningfull survey bootstraping.
}
\details{
\code{draw.bootstrap} takes \code{dat} and draws \code{REP} bootstrap replicates from it.
\code{dat} must be household data where household members correspond to multiple rows with the same household identifier. The data should at least containt the following columns:
\itemize{
  \item Column indicating the sample year;
  \item Column indicating the household ID;
  \item Column containing the household sample weights;
  \item Columns by which population was stratified during the sampling process.
}
A column for the totals in each strat can be included, but is only optional. If it is not included, e.g \code{totals=NULL}, this column will be calculated and added to \code{dat} using \code{strata} and \code{weights}.\cr
The bootstrap replicates are drawn for each survey year (\code{year}) using the function \code{\link[survey]{as.svrepdesign}} from the package \code{survey}.
Afterwards the bootstrap replicates for each household are carried forward from the first year the household enters the survey to all the censecutive years it stays in the survey.\cr
This ensures that the bootstrap replicates follow the same logic as the sampled households, making the bootstrap replicates more comparable to the actual sample units.
}
\examples{
# read in data (must be changed..)
dat <- data.table(read_sas("PATH"))

# create 20 bootstrap replicates using the column "bundesld" as strata
dat_boot <- draw.bootstrap(dat=copy(dat),REP=20,hid="hid",weights="hgew",
                          strata="bundesld",year="jahr")

# do the same with more strata
dat_boot <- draw.bootstrap(dat=copy(dat),REP=20,hid="hid",weights="hgew",
                          strata=c("bundesld","sex","hsize"),year="jahr")

# change column names for bootstrap replicates
dat_boot <- draw.bootstrap(dat=copy(dat),REP=20,hid="hid",weights="hgew",
                          strata=c("bundesld"),year="jahr",boot.names="replicate")

# save bootstrap replicates as .RData
save(dat_boot,file="dat_replicates.RData")
# or .csv-file
write.csv2(dat_boot,file="dat_replicates.csv",row.names=FALSE)

}
\seealso{
\code{\link[data.table]{data.table}} for more information on data.table objects.\cr
\code{\link[survey]{svydesign}} for more information on how to create survey-objects.\cr
\code{\link[survey]{as.svrepdesign}} for more information on how bootstrap replicates are drawn from survey-objects.
}
\author{
Johannes Gussenbauer, Alexander Kowarik, Statistics Austria
}
