% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{draw.bootstrap}
\alias{draw.bootstrap}
\title{Draw bootstrap replicates}
\usage{
draw.bootstrap(dat,REP=1000,hid="DB030",weights="RB050",strata="DB040",
                     year="RB010",totals=NULL,boot.names=NULL)
}
\arguments{
\item{dat}{either data.frame or data.table containing the survey data with rotating panel design.}

\item{REP}{integer indicating the number of bootstrap replicates.}

\item{hid}{character specifying the name of the column in \code{dat} containing the household ID.}

\item{weights}{character specifying the name of the column in \code{dat} containing the sample weights.}

\item{year}{character specifying the name of the column in \code{dat} containing the sample years.}

\item{strata}{character vector specifying the name of the column in \code{dat} by which the population was stratified.
If \code{strata} is a vector stratification will be assumed as the combination of column names contained in \code{strata}.
Setting in addition \code{cluter} not NULL stratification will be assumed on multiple stages, where each additional entry in \code{strata} specifies the stratification variable for the next lower stage. see Details for more information.}

\item{cluster}{character vector specifying cluster in the data. If \code{NULL} household ID is taken es the lowest level cluster.}

\item{totals}{character specifying the name of the column in \code{dat} containing the the totals per strata and/or cluster. Is ONLY optional if \code{cluster} is \code{NULL} or equal \code{hid} and \code{strata} contains one columnname!
Then the households per strata will be calcualted using the \code{weights} argument. If clusters and strata for multiple stages are specified \code{totals} needs to be a vector of \code{length(strata)} specifying the column on \code{dat}
that contain the total number of PSUs at each stage. \code{totals} is interpreted from left the right, meaning that the first argument corresponds to the number of PSUs at the first and the last argument to the number of PSUs at the last stage.}

\item{boot.names}{character indicating the leading string of the column names for each bootstrap replica. If NULL defaults to "w".}

\item{country}{character specifying the name of the column in \code{dat} containing the country name. Is only used if \code{dat} contains data from multiple countries.
In this case the bootstep procedure will be applied on each country seperately. If \code{country=NULL} the household identifier must be unique for each household.}

\item{split}{logical, if TRUE split households are considered using \code{pid}, for more information see Details.}

\item{pid}{column in \code{dat} specifying the personal identifier. This identifier needs to be unique for each person throught the whole data set.}
}
\value{
the survey data with the number of REP bootstrap replicates added as columns.

Returns a data.table containing the original data as well as the number of \code{REP} columns containing the bootstrap replicates for each repetition.\cr
The columns of the bootstrap replicates are by default labeled "w\emph{Number}" where \emph{Number} goes from 1 to \code{REP}.
If the column names of the bootstrap replicates should start with a different character or string the parameter \code{boot.names} can be used.
}
\description{
Draw bootstrap replicates from survey data with rotating panel design.
Survey information, like ID, sample weights, strata and population totals per strata, should be specified to ensure meaningfull survey bootstraping.
}
\details{
\code{draw.bootstrap} takes \code{dat} and draws \code{REP} bootstrap replicates from it.
\code{dat} must be household data where household members correspond to multiple rows with the same household identifier. The data should at least containt the following columns:
\itemize{
  \item Column indicating the sample year;
  \item Column indicating the household ID;
  \item Column containing the household sample weights;
  \item Columns by which population was stratified during the sampling process.
}
For single stage sampling design a column the argument \code{totals} is optional, meaning that a column of the number of PSUs at the first stage does not need to be supplied.
For this case the number of PSUs is calculated and added to \code{dat} using \code{strata} and \code{weights}. By setting \code{cluster} to NULL single stage sampling design is always assumed and
if \code{strata} contains of multiple column names the combination of all those column names will be used for stratification.\cr
In the case of multi stage sampling design the argument \code{totals} needs to be specified and needs to have the same number of arguments as \code{strata}.\cr

If \code{cluster} is \code{NULL} or does not contain the \code{hid} at the last stage \code{hid} it will automatically be used as the final cluster. If, besides \code{hid}, clustering in additional stages is specified the number of column names in
\code{strata} and \code{cluster} (including \code{hid}) must be the same. If for any stage there was no clustering or stratification one can set "1" or "I" for this stage.\cr
For example \code{strata=c("REGION","I"),cluster=c("MUNICIPALITY","HID")} would speficy a 2 stage sampling design where at the first stage the municipalities where drawn stratified by regions
and at the 2nd stage housholds are drawn in each municipality without stratification.\cr

The bootstrap replicates are drawn for each survey year (\code{year}) using the function \code{\link{bootstrap}}.
Afterwards the bootstrap replicates for each household are carried forward from the first year the household enters the survey to all the censecutive years it stays in the survey.\cr
This ensures that the bootstrap replicates follow the same logic as the sampled households, making the bootstrap replicates more comparable to the actual sample units.\cr
If \code{split} ist set to \code{TRUE} and \code{pid} is specified, the bootstrap replicates are carried forward using the personal identifiers instead of the houshold identifier.
This takes into account the issue of a houshold splitting up.
Any person in this new split household will get the same bootstrap replicate as the person that has come from an other household in the survey.
People who enter already existing households will also get the same bootstrap replicate as the other households members had in the previous years.
}
\examples{
library(data.table)

# run on UDB SILC-data

# example for SILC-data for Spain
# dat_es <- fread("path//to//spanish//data.csv")

# approximate Number of clusters if not known
strata <- dat_es[,.(STRATA_sum=sum(RB050[!duplicated(DB030)])),by=list(DB050,RB010)]
strata[,STRATA_ratio:=STRATA_sum/sum(STRATA_sum),by=RB010]
strata[,N.cluster:=random_round(STRATA_ratio*35917),by=RB010]
strata[,N.households:=STRATA_sum/N.cluster]#'
dat_es <- merge(dat_es,strata[,.(DB050,RB010,N.cluster,N.households)],by=c("DB050","RB010"))

dat_boot <- draw.bootstrap(dat=dat_es,REP=250,hid="DB030",weights="RB050",strata=c("DB050","I"),cluster="DB060",
                           year="RB010",totals=c("N.cluster","N.households"),split=TRUE,pid="RB030")

# example for SILC-data for Austria
# dat_at <- fread("path//to//austrian//data.csv")
dat_boot <- draw.bootstrap(dat=dat_at,REP=250,hid="DB030",weights="RB050",strata="DB040",
                           year="RB010",split=TRUE,pid="RB030")


# save bootstrap replicates as .RData
save(dat_boot,file="dat_replicates.RData")
# or .csv-file
write.csv2(dat_boot,file="dat_replicates.csv",row.names=FALSE)

}
\seealso{
\code{\link[data.table]{data.table}} for more information on data.table objects.
}
\author{
Johannes Gussenbauer, Alexander Kowarik, Statistics Austria
}
