% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHHID.R
\name{generate.HHID}
\alias{generate.HHID}
\title{Title}
\usage{
generate.HHID(dat, year = "RB010", pid = "RB030", hid = "db030",
  ID_name = "HHID_new")
}
\arguments{
\item{ID_name}{}
}
\description{
Title
}
\examples{

}
