% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.selection.R
\name{get.selection}
\alias{get.selection}
\title{Title}
\usage{
get.selection(
  dat,
  bootRep = attr(dat_boot, "b.rep"),
  cluster = attr(dat_boot, "cluster"),
  strata = attr(dat_boot, "strata"),
  pid = attr(dat_boot, "pid"),
  hid = attr(dat_boot, "hid")
)
}
\arguments{
\item{dat}{}

\item{bootRep}{}

\item{cluster}{}

\item{strata}{}

\item{pid}{}

\item{hid}{}
}
\description{
Title
}
