% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.surveysd}
\alias{print.surveysd}
\title{Print function for surveysd objects}
\usage{
\method{print}{surveysd}(x, ...)
}
\arguments{
\item{x}{an object of class \code{'surveysd'}}

\item{...}{additonal parameters}
}
\description{
Prints the results of a call to \link{calc.stError}. Shows used variables and
function, number of point estiamtes
as well as properties of the results.
}
