% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalib.R
\name{recalib}
\alias{recalib}
\title{recalib}
\usage{
recalib(dat, hid = "hid", weights = "hgew", b.weights = paste0("w",
  1:1000), year = "jahr", conP.var = c("ksex", "kausl", "al", "erw",
  "pension"), conH.var = c("bundesld", "hsize", "recht"))
}
\arguments{
\item{dat}{either data.frame or data.table containing the sample survey for various years.}

\item{hid}{string specifying the name of the column in \code{dat} containing the household ID.}

\item{weights}{string specifying the name of the column in \code{dat} containing the sample weights.}

\item{b.weights}{string specifying the names of the columns in \code{dat} containing bootstrap weights which should be recalibratet}

\item{year}{string specifying the name of the column in \code{dat} containing the sample years.}

\item{conP.var}{character vector containig person-specific variables for which contingency tables for the population tables are calculatet per year.}

\item{conH.var}{character vector containig household-specific variables for which contingency tables for households are calculatet per year.}

\item{...}{additional arguments passed on to function \code{\link{ipu2}} from the simPop package.}
}
\value{
the survey data with the number of REP bootstrap replicates added as columns.
}
\description{
Recalibrate bootstrap weights by using iterative proportional updating to match population totals on various household and personal levels.
}
\examples{

}
\author{
Johannes Gussenbauer, Statistik Austria
}
